third_party/blink/web_tests/fast/peerconnection/RTCPeerConnection-statsSelector-expected.txt

Tests the RTCPeerConnection stats selector argument.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS pc = new RTCPeerConnection() did not throw exception.
PASS getUserMedia({audio:true, video:true}, gotStream) did not throw exception.
PASS Got a stream.
PASS pc.getStats(statsHandler2, pc.getLocalStreams()[0].getVideoTracks()[0]) did not throw exception.
PASS statsHandler2 was called
PASS result.length is >= 1
PASS timediff is >= 0
PASS successfullyParsed is true

TEST COMPLETE
